fix paths after refactor
